added delay parameter to DumbBot
DumbBot now takes a 4th argument, which is the delay (in seconds) between getting the gamestate and making a move. Setting the delay to 0 causes it to prompt between each move. The lowest actual delay that the server can handle is 0.5, so don't set it to a number below that.